requirements.txt
streamlit
requests
python-dotenv
faiss-cpu
sentence-transformers
numpy
Pillow